Add proper parsing and refactor
So now we've refactored to accept part two of the vod referenced
elsewhere.

I have a few deviations now but it works. I had to deviate because the
person in the VOD has code that doesn't pass the tests, seemingly
missing the parsing of the 'e' in floats and missing a
`self.consume_char()` when parsing digits so the code infinite loops in
the digit parsing.